backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-cov==5.0.0
ruff==0.4.7
mypy==1.10.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
types-passlib==1.7.7.20250602
types-python-jose==3.5.0.20250531
